LATTICE-1590 separate entityset creation for merging
